bytes/iobuf_parser: Safe read truncated utf8
Allows safely reading a fragment of a utf8 string from and iobuf. Naive
truncation of a utf8 string could cut in the middle of a character
sequence.

This adds a utf utility to find the longest valid string under a given
size, and iobuf_parser::read_incomplete_string to read and consume a
size limited utf string from an iobuf