kafka/client: moved decision about request version pre dispatch
The version of the Kafka protocol request may represent a change in the
request semantics. f.e. introduction of incremental fetches. A caller
must be aware of the version to correctly build the request and
handle the response.

Previously the API version was set when the request was dispatched. It
is to late in the request building pipeline. The caller of `dispatch`
method must be able to control the version of request sent.

Signed-off-by: Michał Maślanka <michal@redpanda.com>