ducktape: fix flaky high watermark check in recovery_mode_test
In one test case, one partition of a topic is disabled and the cluster
is restarted to check that the partition remains disabled across
restarts. The test also checks the high watermarks set by producing to
the topic. However, after restart, the cluster must elect new leaders
for the active partitions. The test waits for this, but the node that is
not a part of the partition (there are 4 nodes, partitions are 3x
replicated) may not be updated about the leadership change until
somewhat later than the other nodes. If the describe topic command that
retrieves the high watermark goes to the non-member of the partition, it
can fail saying the partition is leaderless or is in the middle of an
election. This addresses the flakiness by weakening the watermark check
after restart to an eventual check. This should give time for leadership
information to be propagated around the cluster.