serde/json: hide the space trick inside the numeric parser
To parse top-level numbers we're using a trick where we pass a phantom
space into the numeric parser when the JSON parser detects there's no
more input. This forces the numeric parser to finish parsing.
Originally, the trick lived in the JSON parser but maybe it's nicer to
hide it inside the numeric parser.