utils: FIFO logging in retry_chain_context
Previously, the retry_chain_context used a buffer with limited size
which was able to accept new elements until it's full. After it become
full all new log trace messages were discarded.

This commit adds a circular_buffer implementation and uses it instead of
the old buffer. This makes the context evict messages in FIFO order. New
messages are always accepted but they may overwrite some old messages.
The first message can be overwritten partially. In order to avoid
confusion the '[truncated]' prefix is added to the log in this case.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>