app: Propagate correct scheduling group to TS
The initialization logic propagates archival_upload scheduling group to
the cloud_io::io_resources. This is incorrect because this scheduling
group is only used to throttle downloads in the read path. Because of
that the I/O bandwidth is getting limited for archival uploads and not
for TS reads.

This commit fixes this by propagating correct scheduling group. The
archival scheduling group is propagated by the partition_manager.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>