security: Refactor code for parsing license json
The code responsible for parsing the raw license into a license
object has been refactored in a way that is scalable to directly
support more format versions.

Other changes:
- Remove redundant check for negative format version.
By the time this check is performed, version has already been
cast into a uint8_t. On top of it, all non-supported versions
are now covered by the version switch statement.
- Fixed and extended error messages in parsing.
- Zero-initialize license to ensure no field is left uninitialized.