ducktape: fix bad test case with acks=1
`test_relaxed_acks` had one parameterization running with `acks=1`, but
it expected that the consumer would see exactly the set of offsets that
were acked to the producers, even with injected node failures. This is
not always the case:
1. If the leader acks a message but doesn't have time to replicate it
   before failing, and a new leader is elected while the original leader
   is down, the original leader will truncate the acked message from its
   log and no consumer will see it.
2. If the leader acks a message but doesn't have time to replicate or
   flush it to disk before failing, and the leader is reelected when it
   restarts, it may lose the message and no consumer will see it.
Situation 2 was recently seen in CI.

This change removes the `acks=1` case and renames the test, because it's
no longer testing relaxed acks, just default `acks=-1`, with different
levels of caching enabled.