tests/docker: update addr2line SHA
Update the version of the seastar-addr2line tool we use to point to the
tip of our v25.2.x branch, as we've downstreamed the needed fixes, and
have backtrace decoding optimizations in there that we want to use.

While we are at it, DRY the script and add a small doc.

CORE-12786