rpk profile create: fix --from-cloud flag
Cobra has a particular restriction when using
the NoOptDefault option, once selected, the user
cannot use the syntax '--flag value', instead,
now it has to always use '--flag=value'.

This commit fixes that by manually parsing the
command arguments and doing the validation
before the execution.

Also, fixes a bug where a user with an empty
profile name (manually modified the rpk.yaml)
would not be able to create a new profile without
renaming the empty one. Is a very rare edge case
but it doesn't hurt to fix this.