c/a: Re-allow retries in ntp_archiver::upload_segment
A previous PR[1] refactored upload_segment and in so doing added
retry_strategy::disallow to the retry_chain_node governing the upload.

As a result, we saw an uptick in
- 'cloud_storage_failed_uploads'
- 'cloud_storage_bytes_sent'
- 'io_queue_total_read_bytes'

Along with an increased incidence of 'backoff quota exceeded' logs from
cloud_io.

Specifically cloud_io::remote::upload_stream increments
'cloud_storage_failed_uploads' in a number of failure cases, but along with
the increased frequency of the backoff quota log line, it's likely bordering
on a certainty that the uptick in failures is a result of rtc node retries
being exhausted as a result of transient errors that would previously be
masked by retry logic inside cloud_io.

For now, we should return to the default retry strategy (exponential backoff)
and assess the viability of offloading retries to the archival loop at a
later time.

[1] https://github.com/redpanda-data/redpanda/pull/25951

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>