`storage`: pass `seg->size_before()` as `segment_size` hint to `appender`
During compaction, it is guaranteed that the size of the `segment` on disk
after a de-duplication operation is less than the size of the `segment`
before the operation. There is a handy tool within the `segment_appender`
by which you can pass a `segment_size` as a hint to not overly `fallocate()`
more space than is necessary on disk (as doing so is wasteful and leads to
greater fragmentation, since we will `ftruncate()` the file to its rightful
size after writing to it).

In the pathological case where we have many small `segment`s on disk being
continually compacted and re-compacted, the default `segment_fallocation_step`
of `32_MiB` may be _much_ more space than is necessary (we have seen instances
where `segment`s are on the order of hundreds of `KiB`, and certainly less than
tens of `MiB`.

We were not making use of this hint to the `segment_appender` in the
compaction routines. This lead to us previously adjusting the
`segment_fallocation_step` for a user experiencing the pathological case of
_many_ small `segment`s from `32_MiB` down to `2_MiB`, which yielded great
results for their p99 latency (which was incredibly high due to time spent
searching for memory in the XFS filesystem during an `fallocate()` call,
due to the aforementioned fragmentation which we have great reason to believe
is entirely due to the high `segment` count/low `segment` file size).

Optimize this code path for the pathological case by providing the
pre-compacted `segment` size as a hint, since we know that there is no
possibility for the `segment` size to be higher post compaction.