archival: remove upload retry limit for segments
In a previous PR[1] we began to rely on the archiver loop to retry, and
moved away from relying on `cloud_io::remote` for retries in two ways:
1. setting an explicit `disallow` retry policy on the retry node passed
   to the remote, and
2. setting the `max_retries` passed to `remote::upload_segment()` to 1.

In practice, we saw that _not_ relying on the remote resulted in an
uptick in the `vectorized_cloud_storage_failed_uploads` metric, which is
monitored and alerted on. In [2] we reverted #1, but didn't notice #2.
This commit reverts #2.

[1] https://github.com/redpanda-data/redpanda/pull/25951
[2] https://github.com/redpanda-data/redpanda/pull/26969