base: introduce format_to helper
To make creating formatters less verbose, we're creating a custom
extension method that is much simpler to make types formattable using.

The alternatives seem to be a `macro` or the very verbose method of
implementing a fmt::formatter for each type we have. Having this
convention of `format_to` instead makes it easier for folks to not write
`operator<<` instead and use this method instead. Future versions of
libfmt don't support using `operator<<` anyways.