chunked_vector: remove fragmented vector impl
The final change in this series transitioning from fragmented_vector to
chunked_vector.

Previously the `fragmented_vector` was a dual implementation that could
be used as a chunked vector or a fragmented vector depending on the
template arguments provided, either a valid `fragment_size` for frag
vector or std::dynamic_extent for chunked vector.

This change removes the duality and simplifies the implementation by
removing the `fragment_size` template argument. The `chunked_vector`
now becomes the sole implementation for chunked vectors, and it is no
longer capable of being used as a fragmented vector.

Comments are updated where necessary to reflect the new state of the
code.

Also updates type_traits.h to accommodate this change.

Fixes CORE-12882.