ct: Refactor data_plane_impl
Make 'impl' class use the 'ssx::sharded_service_container'. Make
pipelines and their components to be sharded services.

The motivation here is that we want some cross-shard functionality in L0
(aggregated upload per broker and not per shard) which will require us
to create a peering sharded service. But with the current approach this
is impossible. The whole 'impl' is a monolitic object which is 'sharded'
through a wrapper (state_accessors).

This commit also changes 'state_accessor' to be non-owning. It stores a
ref to the 'data_plane_api' object. In the current form its only purpose
is to pass state around without introducing dependencies.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>