batch/compression: don't allow compressing with model::compression::none
There are a few things here, one now we require that batches passed into
compress_batch are marked as uncompressed (callers updated), otherwise
if you tried to *change* compression type this method would mess up the
bit flags for compression in the header (thankfully we never do this,
but we should shore up the API).

Secondly we disallow calling `compress_batch` if you want the batch to
be uncompressed.

Lastly we throw instead of vassert, as this doesn't seem worthy of a
vassert.