datalake: Allow replacing dots in table names
This defines a configuration property `iceberg_topic_name_dot_replacement`
that, when set, will replace .'s in table names when a Kafka topic name
is translated into the corresponding Iceberg table name (or DLQ table name).

The replacement is done using the current value of the setting, so to
preserve the link between topic and table, the setting should not be
changed once there is an Iceberg topic whose name contains dots. It'd be
nice to copy the cluster configuration into topic configuration when an
Iceberg topic is created or when Iceberg mode is set for a topic, so the
link is preserved even if the cluster configuration is changed, but
that's a far more invasive change. This workaround may suffice.