rptest: make llvm-symbolizer available in test env
In order for ASAN, UBSAN, and other sanitizers to produce symbolized
stack traces, the llvm-symbolizer binary needs to be available in the
PATH. This change makes it available in the test environment, both
Docker and bare-metal (CDT).

We were already installing this using llvm.sh, so we simply need to
symlink it into /usr/local/bin.

Fixes CORE-13191.