rpk.py: trim exception string
The RPK exception sometimes embeds the entire stdout/stderr of the
command that was run, which can be very large, especially the stderr
since we turn on RPK debugging so there is verbose output about every
request.

These exceptions may casually be logged to the console, or end up in a
report, where this amount of info isn't useful and causes other issues.

Instead, include only the last two lines of stdout/stderr in the
exception __str__, but keep the full text in the exception object for
those who want to dig into it (i.e., don't change the contract of
RpkException, except its __str__).

Furthermore, in failure cases, log the full stdout/stderr to the test
log at debug level, which is probably where folks would want it.

Adjusts one test that was looking for AUTHORIZATION_FAILED in the
exception __str__ form to look in stderr instead.

Fixes CORE-13195.