cloud_topics: add manager
Introduces a top-level manager component for cloud topics. The idea is
for the manager to coordinate work and manage state that is inherently
global to cloud topics. The current user is the level zero garbage
collection process that need only have a location in the cluster to run,
so it attaches to the active manager instance.

The manager itself as well as the level zero GC process are stateless.
Because of the stateless nature, an entirely new internal, dedicated
topic/partition is avoided and instead the manager attaches to
partition-0 of the L1 domains topic. Once the manager or its dependents
become stateful or for whatever reason we want to avoid attaching to the
L1 domains NTP, then the attach target NTP in the mananger can be updated.

For now using the L1 domains topic coordinates nicely with the
dependencies of L0 GC process because this topic must exist before any
L0 data ever becomes eligible for GC.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>