c/partition_notifier: added option to notify on current partitions
Added an option to immediately notify on partitions that are currently
being managed by the partition manager.

The notification may be registered in the
`cluster::partition_change_notifier` after some partitions are already
present in the `partition_manager` in this case the notification would
be missed. (good example is the `controller` partition which is created
at the very beginning of the startup process).

Added caller an option to be immediately notified about all existing
partitions in the manager when registering notification.

Signed-off-by: Michał Maślanka <michal@redpanda.com>