`ct/l1`: add `offset_interval_set::covers()`
For determining if an input interval is covered by an interval present in
the `interval_offset_map`.

This will be used in the future for determining which `extent`s in
a partition are cleanly compacted or not. For example, computing
the dirty ratio of a log could be written as:

```
double get_dirty_ratio(const partition_state& state){
    size_t total_size{0};
    size_t dirty_size{0};
    const auto& cleaned_ranges = state.compaction_state->cleaned_ranges;
    for (const auto& extent : state.extents) {
        total_size += extent.len;
        auto b = extent.base_offset;
        auto e = extent.last_offset;
        if (!cleaned_ranges.covers(b, e)) {
            dirty_size += extent.len;
        }
    }
    return total_size == 0 ? 0.0
                           : static_cast<double>(dirty_size)
                               / static_cast<double>(total_size);
}
```