direct_consumer/fetcher: deduplication fix
Fetcher allows for concurrent updates to its assignments
- assignments are held in
  chunked_hash_map<topic, chunked_hash_map<partition_id, fetcher_state>>
- assignments may be updated on any fiber by modifying this assignments
  map

Fetcher state is an intrusive list element

The fetcher runs in a loop fiber as follows:
- partitions = collect_partitions()
- request = make_request(partitions)
- fetch_response = dispatch_fetch(request)
- update_assignments(fetch_response)

Where collect partitions builds an intrusive list of partition
to include in the next fetch request from the assignments map
and make_request uses that intrusive list to serialize a request

Currently, this approach has two problems.
1. fetcher_state is an intrusive list element but has no custom move
   semantics.
   - naiive moves of intrusive list elements do not preserve their
     intrusive lists
   - removals from chunked_hash_map can move the contained elements
   - can surface as an unexpected reordering of the partitions to
     include list used by make_request
     - the symptom of this was that partitions would be missed in some
       fetch requests, thus failing to inform their broker of updated
       fetch offsets. This lead to duplicated fetched records
2. iterators to intrusive lists are invalidated by the destruction of
   their backing intrusive list element
   - this means that a sufficiently unlucky unassignment could destruct
     the element that the fetcher loop is currently iterating on causing
     a SIGSEV

The resolution used in this commit is to stop using intrusive list for
the incremental state created by the fetcher loop fiber.

Instead, the incremental state is now held in a chunked_vector which
owns a copy of the fetcher state as it was at the beginning of the loop.

Now that the fetcher loop owns the memory on which it iterates, the
backing list can be more easily shown NOT to mutate while the fetcher
iterates.

This will prevent both of the above issues.