`compaction`: rename `{backward/forward}_iteration()`
These names were a poor choice as they describe implementation details
(which may not necessarily be true) and not the actual purpose of
the functions.

Rename them to `map_building_iteration()` and `deduplication_iteration()`
respectively to be more descriptive about their functionality (and not
perscribe any requirements to their implementation).