ct: Implement inactive epoch estimation
The ctp_stm implements get_inactive_epoch method which creates a log
reader and scans the partition to compute inactive epoch. This commit
adds an alternative lightweight mechanism that can be used to estimate
the value of the inactive epoch. The estimated value is guaranteed to
never overestimate the inactive epoch.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>