ct/l1: fix reader performance with compacted L1 data
Currently when there are gaps in the log at the end of object, we will
make a metastore request for every offset at the end of the object.
Prevent this from being an issue by having the metastore return the
last offset available in the object, so we can skip directly to the next
object in the log.