storage/ntp_config: ensure features are disabled for cloud topics
Tiered storage and cloud topics are mutually exclusive properties. We
should probably have validation at some level for this stuff in the
alter configs path. Some other features like RRR and Iceberg are not
supported in cloud topics for the immediate feature, so let's just
hard disable them for now.