k/metadata: fixed returning incorrect leader epoch
Since version 9 the `MetadataResponse` contains an information about
partition leader epoch. If present the epoch value is considered by the
clients as valid i.e. they use the leader epoch value to fence stale
metadata updates. In some rare cases f.e. when some nodes are in
recovery mode while other are not Redpanda leaders table may not yet
contain the leader epoch information. In this case `MetadataResponse`
should return an invalid leader epoch. This way the clients will not use
the epoch to fence the partition update.

Together with implementation of TopicIDs support we upgraded max handled
`MetadataRequest` version from `8` to `12`. This way clients started to
rely on the leader epoch information included in `MetadataResponse`.

Previously when no leader epoch was present in partition leaders cache
the metadata request handler always returned first replica as a leader
and `0` as a value of leader epoch. This was incorrect as it forced
clients to fence the partition metadata update. As a result of the
update fencing the partition was not included in the consecutive fetch
requests leading to a situation in which consumers were stuck for time
required to refresh metadata again.

Described issue caused some of the tests (the one using Java based Kafka
clients) to fail intermittently.

Fixes: #CORE-9955

Signed-off-by: Michał Maślanka <michal@redpanda.com>