ct/retention: introduce housekeeper
Ah our wonderful housekeeper for cloud topics who ensures that each
locally attached CTP is properly keeping the correct amount of data
according to the topic's retention policies. We'll have this
managed/attached using the cloud_topics manager and it will coordinate
between L1 and L0 to compute the start offset.

Note that intentionally are a little lax with bytes based retention. We
don't account for L0 data, and we allow going over the limit a bit (this
is implemented in the metastore, but still). Given cloud storage not
being a limited resource in the same way as local disk, I think it's OK
to be lax here, as it simplifies the implementation.