rm_stm: async loop over intrusive list fix
It is dangerous to suspend while looping over intrusive lists because
destruction of the intrusive list element can invalidate an iterator on
the intrusive list.

Here, the resolution is to snapshot the producer_identities which need
to be cancelled into an owning collection before crossing a suspension
point. The snapshotted producer identities will subsequently be used to
expire transactions, which is an asynchronous operation.