random_generators: allow random seeding
Add a constructor for rng that always uses a random seed, and a
helper function to get an rng with a random seed. This allows
select code that cannot cope with a fixed seed to explicitly request
a random seed.

The primary use case is tests that use randomness to generate a unique
test directory. Probably a bad practice, but for now we just cope with
that.