ct/l0/ctp_stm: don't wastefully replicate a no-op LRO advance cmd
This actually fixes a real bug too, if reconcilation tries to reconcile
LRO `kafka::prev_offset(0)` which in `kafka::offset::min()` - this will
throw translation errors when the offset is applied to the STM. The fix
is to just not replicate offsets if it won't do anything, this prevents
this case. All other cases should be a bug.