pandaproxy: relax client lock from mutex to rwlock
This client_lock is only used for coordinating the stopping of clients.
It is sufficient for this to be a read write lock instead of a mutex,
since that still keeps the stopping coordination safe, but allows for
multiple concurrent requests on the request handling path.

This mutex currently limits performance to 1 request/user when http
basic authentication is used. Turning the mutex into a rwlock should
significantly improve parallelism in this case.