`l1/compaction`: add `compaction_info_and_timestamp` to `meta`
To track `compaction_info` sampled from the `metastore` as well as
the timestamp at when it was sampled.

Also adds the `compaction_log_queue` typedef, which is a priority queue
holding pointers to `log_compaction_meta`s which require compaction.
The idea here is that a `log_compaction_meta*` with `info_and_ts` set
and requiring compaction should be pushed into a `priority_queue` held
by a scheduler/manager in order to allow `compaction_worker`s to pull
jobs from the global queue and complete them on their shard.