ct/housekeeper: Sync L0 start offset to L1
L0 is the source of truth for the start offset, even though the start
offset is computed based on the retention config and information in the
L1 metastore. This is because a Delete Records request may advance the
start offset outside of the housekeeping retention mechanism, and it
means that the L0 start offset needs to be propagated back to the L1
metastore. This change makes the housekeeper perform the sync at the end
of its run.