random: fix static initialization order fiasco
The seeding mode was subject to the static initialization order fiasco
as a global variable that could be accesses by other globals before it
was initialized. This happens in practice in role_store_bench and
probably other places.

The result was that random mode was always selected, at least for the
global rng object. This could affect any objects created until the
next reseeding event.

Move seeding mode initialization to function-local static to avoid
static initialization order fiasco. Also add logging about the env
var values in bench-wrapper and a test for the global rng values (
though this test did not fail before this fix, probably because
it's a "fiasco" and the order cannot be predicted).