tuner: Same all-nics-same-mode checker
This adds a "tuner" which checks that all to be tuned NICs share the
same mode and interrupt config.

This currently isn't enforced or checked so chaos can ensure if NICs are
different.

The implementation is a bit odd but that's due how the tuner and the
different net subtuners currently work.

In the future I want to remove all the sub tuners and just have a single
tuner which would make this more natural.