tuner: Write net tuner config file
Adds a tuner which writes the in use tuner config (cpusets) to a file in
/var/run.

This file can then be used by rpk:start to make Redpanda start with the
correct cpuset.

This is the preferred alternative to rpk:start doing the auto detection
logic as that won't work in k8s or similar.

We use /var/run as that doesn't get cleaned by things like tmpfiles.d.
It requires root but the tuner needs running as root anyway.