raft: Propagate expected term through replicate_options
This commit adds a plumbing for the expected term. Previously, the
expected term was passed directly into 'raft::replicate' and
'raft::replicate_in_stages' (2x overloads of each method existed).

This commit adds expected_term to replicate_options.
The field is supposed to be used to replace the 'replicate' method
overloads that take expected term.

Since now we have expected_term field in the replicate_options we no
longer need to pass this value directly. This commit removes overloads
that accept exected term and propagates the value all the way up to the
batcher and updates use cases.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>