iobuf: optimize comparison more
Turns out if the first byte is different memcmp is slower than a byte
for byte comparison. So what we do is optimize small iobufs and change
how we do the comparison, instead we check. This is a tradeoff we can
see in the bench for the best/worse case of how many bytes we compare