admin: refactor with generic connection_collector
Introduce an abstract connection_collector that can be used to implement
the collection of connections in a generic way based on whether ordering
is used or not.

In the (existing) unordered case, the collector collects the first
`limit` elements into a chunked_vector and ignores all later elements.

In the ordered case, the collector will collect up to `limit` elements
into a `bounded_chunked_priority_queue`.

A local collector is used per-shard. This is aware of the remaining
number of items to collect in the unordered case to ensure that at most
`limit` items are collected globally.

A global collector is then used to "merge in" the shard-local results.

This collector abstraction helps reuse code across the ordered and
unordered use cases, while remaining memory efficient (only using up to
O(limit) memory) in either case.