tuner: Empty tuner config file instead of deleting it
More k8s problems. k8s FileOrCreate mode creates an empty file if the
file doesn't exist on the hostPath.

This is already handled by rpk start. However there is another problem
in that this file/mount doesn't seem to be deletable so the MQ mode
fails as it tries to delete the empty file.

Change it such that in MQ we just empty it instead (if it's not
already).