dt: capture the crash_reports dir on redpanda nodes
The crash_reports directory contains information about why the redpanda
node crashed. In some cases we have seen ducktape failures where there
is a log line stating that the crash has been successfully recorded, yet
the stderr logs do not contain the stacktrace. In this case, capturing
the crash_reports directory should give us a way to see the cause of the
crash.