`compaction`: rename `is_compactible_control_batch` to `is_removable...`
"Compactible" is now a term that suggests a record may be removed due to
de-duplication (i.e a record with the same key and a higher offset in the log
exists).

Rename this function to "is_removable_control_batch()", since it returns `true`
for control batches that can be immediately removed without any consideration of
the rules of de-duplication.