dm: prevent topic from being active at the same time in two clusters
For correctness the same topic must never be instantiated in the source
and target cluster at the same time. (this may lead to issues as
archivers on the cluster may try to modify the topic state at the same
time.).

In order to make sure the topic is never active on two cluste this
commit changes the migration logic. In the outbound migrations the topic
is unmounted in `executing` state and then deleted in `cut_over` state.
Changes in inboud migration involve deferring the topic creation to
`cut_over` state.

Signed-off-by: Michał Maślanka <michal@redpanda.com>