cluster/epoch_service: prevent bypassing shard0 cached epoch
We need to enforce that we never bypass shard0 for the epoch. Calling
get_current_epoch from any shard can cause another shard to race head of
the current cached value. This is problematic if leadership changes and
the ahead shard re-asks shard0 for the leader ID, because then we see an
epoch regress, which breaks an invariant we want to hold.

Instead only allow asking shard0 for the epoch in a manner that ensures
its cached value is updated as well.