rpk shadow update: allow easy password change
the password is an input_only field, for outputs
we have the `password_set` boolean. Now we use it
to display a '<redacted>' string to the user can
just replace that string to set a new password.

Before this, the password will just not be in the
configuration shown during update, making it
difficult for users to update passwords (having
to type the whole 'password: ...' field.