dt: fix cl rapid toggling test
During the toggling, the topic syncer sometimes gets stuck.
Under the previous setup, the timeout time was encountered before the
toggling had the time to finish. Change the order to do all the toggling
and only then try to verify if all the topics will reach the expected
state.

Args were reduced from 200 to 100 (still a big number) to save time.
This makes the test about a minute faster, when tested locally.