type-check: skip the docker build in GHA
We already build it using the build-push-action, so no need to do it
again (and the cache is not effective as different drivers are used).

We also need to align the docker image name in GHA and the type-check.sh
script, otherwise we won't find it (the latter used "checking" vs
"check" in GHA).