s/appender: do not append to chunk which is being dma_written
This commit introduces a logic that prevents the head chunk from being
appended if it is dispatched i.e. its content is being DMA written.

Previously the head chunk buffer might have been passed to the
`file::dma_write_call` but appends could still write to that chunk. This
caused a problem which manifested on some filesystems and some hardware
RAID controllers. The problem manifested as corruption in the last page
of a file on disk.

This PR prevents the head being written to disk from being concurrently
appended. The mechanism used here is copying the reminder which wasn't
flushed and spans beyond the last page aligned address to a new chunk.
The reminder is copied only if the current head write was dispatched.
After copying the chunk writer is requested to recycle the chunk and
return it to the cache.

Signed-off-by: Michał Maślanka <michal@redpanda.com>