lsm/core: add compression routines
This is very similar to the model/batch_compression routines, but in
order to keep the LSM implementation standalone, we'll reimplement a
similar strategy for compression and decompression of payloads.

The write path is all using `iobuf`, but since reads need random access
they use `ioarray` for decompression. For now, we don't implement any kind
of optimized compression for `ioarray`, but we can save the conversion
to `iobuf` for decompression in a followup.