lsm/core: introduce internal options
This is the meat of the configuration for the LSM tree and will be
plumbed around in a few layers. Right now we seperate the options that
will be provided internally vs externally, so we can optimize for
different things. Interally we likely want to precompute state or have
invariants in place (such as with bloom filter size), but we don't want
to place that burden/knowledge on users, so we will have a seperate set
of options that will be passed into the subsystem, and these internal
options will be created from them.