c/migrations: do not assert when building group partition map
When migrating consumer groups migration backend builds a map that
associate consumer groups with `__consumer_offsets` topic partitions.
The map is built when the migration is reconciled. Previously the race
condition between application of state in migration and topic table
could lead to a situation in which a `__consumer_offsets` partition is
not yet present in partition manager. This is when assertion was
triggered.

Added retry with small backoff to wait for the partition to appear
instead of asserting and crashing Redpanda application.

Fixes: CORE-14421

Signed-off-by: Michał Maślanka <michal@redpanda.com>