rptest: set the full scrub internal low in cleanup
In the test cleanup, we wait for a cloud stoarge scrub to happen in
order to identify any anomalies. To make this happen in a timely
manner, we have to do something. Currently we reset the scrub metadata
which actually ends up triggering a "first scrub" right away, ignoring
the partial/full scrub interval.

However, if a leadership change happens at a critical moment, the
metadata ends up being populated again, so the "full scrub" interval
will apply is 10 minutes. This will time out the test.

To fix set the full scrub interval down to 10 seconds so the scrubs
happen "soon" even in this case. Because we reset the scrubbing
metadata, the next scrub will effecitvely be a full scrub (since the
last scrubbed offset will be forgotten), so that is the interval that
applies.

We don't set it super low in order not to do a ton of full scrubs
during this teardown phase.

Fixes [CORE-14424].