storage/tests: fail fast(er) in segment ms test
This test had a few failures with test timeouts (300s) due to timeouts
during produce. I think they were caused by bad disks we ran the tests
on for a while.

Anyway, we can do better than wait for 300 seconds for test to fail
(never if run outside the test framework). Always mark the "async
produce" as completed so that the test can continue with other
assertions including checking that "produce" tasks succeeded.