cluster_link/model: change default max partition bytes
Prior to this commit the default max partition bytes was 1MiB. The
default min fetch bytes was 5MiB. And the max wait ms was 500ms.

In cases where only one partition on the shard was being produced to
these defaults limited our consumer throughput to 2MiB/s. This is since
we'd only be able to read 1MiB every 500ms from the one partition with
new data.

Hence this commit sets max partition bytes to equal min fetch bytes to
prevent this issue.