sec/oidc: do not wait for EOF on OIDC requests
The typical TLS shutdown process consists of the client sending a
close_notify to the server, then the server closing the connection in
response to that.

However, many servers are non-compliant and they just drop the
connection on close_notify instead of doing an orderly shutdown on it.

We see this in particular in our default-configured OIDC server, which
is showing up as flakiness in the OIDCTlsTest's.

It is in-spec to not wait for EOF on shutdown, in fact, a recent
upstream seastar change will make this the default behaviour.

So this commit implements a targeted, easily-backportable
test-flakiness-fix by skipping waiting for EOF for requests to the OIDC
server.

Ref: https://datatracker.ietf.org/doc/html/rfc8446

>   Both parties need not wait to receive a
>   "close_notify" alert before closing their read side of the
>   connection, though doing so would introduce the possibility of
>   truncation.

and because we are using TCP we have that

>   Note: It is assumed that closing the write side of a connection
>   reliably delivers pending data before destroying the transport.