rptest: remove "family" from MetricsSample
This was a confusing concept had little utility. It is either equal to
the metric name, or the name with the suffix like "_bucket" or "_sum"
or "_total" removed.

It had only a single, probably inadvertent use in datalake_e2e_test.py
where it was logged instead of the metric name, change that to name.

Update the documentation accordingly, and also document MetricsSample
fields.